GouelFront/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
black==23.12.0
blinker==1.7.0
cachelib==0.10.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
Flask==3.0.0
Flask-JSON==0.4.0
Flask-Session==0.5.0
frozenlist==1.4.1
idna==3.6
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pycparser==2.21
pyOpenSSL==23.3.0
python-dotenv==1.0.0
redis==5.0.1
redispy==3.0.0
requests==2.31.0
urllib3==2.1.0
Werkzeug==3.0.1
yarl==1.9.4
